backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
python-decouple==3.8
psycopg2-binary==2.9.10
drf-yasg==1.21.9
# Pillow==11.1.0  # Установим позже после настройки system libraries
